backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.10.1
python-multipart==0.0.20
python-dotenv==1.0.1
cors==1.0.1